refactor(doctor): split environment.rs into mod + audits
475-line environment.rs split: mod.rs keeps the main orchestrator and
strict_status helper; audits.rs holds the 5 self-contained checks
(home_hygiene, path_conflict, captured_env, manager_conflict,
python_base).
